bradax-broker/requirements.txt
# FastAPI e servidor ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Configuração e validação
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging estruturado
structlog==23.2.0

# JWT e autenticação
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4

# gRPC
grpcio==1.59.3
grpcio-tools==1.59.3

# Banco de dados
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Redis
redis[hiredis]==5.0.1

# HTTP client para LLMs
httpx==0.25.2
aiohttp==3.9.1

# Métricas e monitoramento  
prometheus-client==0.19.0
psutil==5.9.6

# Vector database clients
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
